- added multi selection for samples - samples can be provided to be include in song - individual song can be started fresh (without preceiding MAS song creation), new song dir will be created - add edit on song title in creative mode - default images shown when non are created - light + full chat history is kept - created seperate visualization component
